Add support for compiling on 32-bit systems
A few explicit casts are needed in order to get the
complete buildbarn code to compile on 32-bit systems.

This makes it possible to e.g compile and execute buildbarn
on raspberry pi:s running on 32-bit kernels.

+++ pkg/runner/processtablecleaning/system_process_table_linux.go
@@ -49,7 +49,7 @@ func (pt systemProcessTable) GetProcesses() ([]Process, error) {
 		processes = append(processes, Process{
 			ProcessID:    int(pid),
 			UserID:       int(stat.Uid),
-			CreationTime: time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec),
+			CreationTime: time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec)),
 		})
 	}
 	return processes, nil
